# backend/docs/swagger.yaml
definitions:
  dto.CancellationDTO:
    properties:
      cancellation_id:
        type: string
      datetime:
        type: string
      product_id:
        type: string
      product_name:
        type: string
      quantity:
        type: number
    type: object
  dto.PriceDTO:
    properties:
      date_end:
        type: string
      date_start:
        type: string
      discount:
        type: number
      full_price:
        type: number
      price_id:
        type: string
      product_name:
        type: string
      total_price:
        type: number
    type: object
...
  models.Cancellation:
    properties:
      cancellation_id:
        type: string
      datetime:
        type: string
      product:
        $ref: '#/definitions/models.Product'
      product_id:
        type: string
      quantity:
        type: number
    type: object
  models.Price:
    properties:
      date_end:
        type: string
      date_start:
        type: string
      discount:
        type: number
      full_price:
        type: number
      price_id:
        type: string
      product:
        $ref: '#/definitions/models.Product'
      product_id:
        type: string
      total_price:
        type: number
    type: object
  models.Product:
    properties:
      manufacturer:
        type: string
      name:
        type: string
      product_id:
        type: string
      type:
        $ref: '#/definitions/models.ProductType'
      type_id:
        type: string
      volume:
        type: number
      weight:
        type: number
    type: object
...
  models.ProductType:
    properties:
      for_adult:
        type: boolean
      name:
        type: string
      type_id:
        type: string
    type: object
...
  response.Response:
    properties:
      error:
        description: Error if response is error
        type: string
      status:
        description: Response status
        type: string
    type: object
info:
  contact: {}
paths:
  /cancellations:
    get:
      description: Returns a list of cancellations filtered by product and date range.
      parameters:
      - description: Product ID
        format: uuid
        in: query
        name: product_id
        type: string
      - description: Start date (YYYY-MM-DD)
        in: query
        name: date_from
        type: string
      - description: End date (YYYY-MM-DD)
        in: query
        name: date_to
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/dto.CancellationDTO'
            type: array
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.Response'
      summary: Get cancellations
      tags:
      - cancellations
    post:
      consumes:
      - application/json
      description: Creates a new cancellation.
      parameters:
      - description: Cancellation
        in: body
        name: cancellation
        required: true
        schema:
          $ref: '#/definitions/models.Cancellation'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            items:
              $ref: '#/definitions/dto.CancellationDTO'
            type: array
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.Response'
      summary: Create cancellations
      tags:
      - cancellations
  /cancellations/{id}:
    delete:
      consumes:
      - application/json
      description: Deletes a cancellation by ID.
      parameters:
      - description: Cancellation ID
        format: uuid
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.Response'
      summary: Delete cancellation
      tags:
      - cancellations
    put:
      consumes:
      - application/json
      description: Updates an existing cancellation.
      parameters:
      - description: Cancellation ID
        format: uuid
        in: path
        name: id
        required: true
        type: string
      - description: Cancellation
        in: body
        name: cancellation
        required: true
        schema:
          $ref: '#/definitions/models.Cancellation'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.Response'
      summary: Update cancellation
      tags:
      - cancellations
  /prices:
    get:
      description: Returns a list of prices filtered by product type and date range.
      parameters:
      - description: Product type ID
        format: uuid
        in: query
        name: type_id
        type: string
      - description: Start date (YYYY-MM-DD)
        in: query
        name: date_from
        type: string
      - description: End date (YYYY-MM-DD)
        in: query
        name: date_to
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/dto.PriceDTO'
            type: array
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.Response'
      summary: Get prices
      tags:
      - prices
    post:
      consumes:
      - application/json
      description: Creates a new price.
      parameters:
      - description: Price
        in: body
        name: price
        required: true
        schema:
          $ref: '#/definitions/models.Price'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/models.Price'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.Response'
      summary: Create price
      tags:
      - prices
  /prices/{id}:
    delete:
      description: Deletes a price by ID.
      parameters:
      - description: Price ID
        format: uuid
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.Response'
      summary: Delete price
      tags:
      - prices
    put:
      consumes:
      - application/json
      description: Updates an existing price.
      parameters:
      - description: Price ID
        format: uuid
        in: path
        name: id
        required: true
        type: string
      - description: Price
        in: body
        name: price
        required: true
        schema:
          $ref: '#/definitions/models.Price'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.Response'
      summary: Update price
      tags:
      - prices
...
swagger: "2.0"